Move to internal pkg. Change exposed struct name
